backend/requirements.txt
# Django Backend Requirements for GenFree Network

# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary
dj-database-url==2.1.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-allauth==0.56.1
cryptography>=3.4.8

# Real-time features
channels==4.0.0
channels-redis==4.2.0
redis==4.6.0

# File handling & Media
Pillow
django-storages==1.14.2
boto3==1.29.0

# Payment processing
requests==2.31.0
python-decouple==3.8

# Background tasks
celery==5.3.4
django-celery-beat==2.5.0

# API documentation
drf-spectacular==0.26.5

# Utilities
python-slugify==8.0.1
django-extensions==3.2.3
django-filter==23.3

# Development
django-debug-toolbar==4.2.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Monitoring
sentry-sdk==1.38.0